# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Landa Travels conversational planning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me/handovers:
    get:
      operationId: listMyHandovers
      tags: [traveller]
      summary: List the current traveller's advisor handovers
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Privacy-minimised handovers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdvisorHandover" }
        "401": { description: Authentication required }
  /me/handovers/{id}:
    get:
      operationId: getMyHandover
      tags: [traveller]
      summary: Get one of the current traveller's advisor handovers
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Advisor handover
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdvisorHandover" }
        "401": { description: Authentication required }
        "404": { description: Handover not found }
...
  /me/profile:
    get:
      operationId: getCurrentProfile
      tags: [traveller]
      security: [{ bearerAuth: [] }]
      summary: Get the current traveller profile
      responses:
        "200":
          description: Current profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravellerProfile"
    patch:
      operationId: updateCurrentProfile
      tags: [traveller]
      security: [{ bearerAuth: [] }]
      summary: Update the current traveller profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravellerProfile"
  /trips:
    get:
      operationId: listSavedTrips
      tags: [traveller]
      security: [{ bearerAuth: [] }]
      summary: List saved trips for the current traveller
      responses:
        "200":
          description: Saved trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedTrip"
    post:
      operationId: createSavedTrip
      tags: [traveller]
      security: [{ bearerAuth: [] }]
      summary: Save a trip plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedTripInput"
      responses:
        "201":
          description: Saved trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedTrip"
  /trips/{id}:
    patch:
      operationId: updateSavedTrip
      tags: [traveller]
      security: [{ bearerAuth: [] }]
      summary: Update a saved trip
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedTripUpdate"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedTrip"
  /bookings:
    get:
      operationId: listBookings
      tags: [traveller]
      security: [{ bearerAuth: [] }]
      summary: List requested or demo booking records
      responses:
        "200":
          description: Booking records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [traveller]
      security: [{ bearerAuth: [] }]
      summary: Create a booking request record
      description: Creates a demo/requested record only. It never charges or confirms inventory.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /rewards:
    get:
      operationId: getRewardSummary
      tags: [traveller]
      security: [{ bearerAuth: [] }]
      summary: Get sustainability points and evidence
      responses:
        "200":
          description: Reward summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardSummary"
  /contact:
    post:
      operationId: submitContactRequest
      tags: [contact]
      summary: Submit a public contact request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Contact request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
...
components:
...
  schemas:
...
    AdvisorHandover:
      type: object
      required: [handoverId, savedTripId, status, summary, travellerReply, assignedAdvisorId, assignedAdvisorName, notificationStatus, createdAt, updatedAt]
      properties:
        handoverId: { type: string }
        savedTripId: { type: ["integer", "null"] }
        replyNotificationStatus:
          type: string
          enum: [not_requested, pending, accepted, failed, not_configured]
        status:
          type: string
          enum: [requested, assigned, replied, closed]
        summary: { type: string }
        travellerReply: { type: ["string", "null"] }
        assignedAdvisorId: { type: ["string", "null"] }
        assignedAdvisorName: { type: ["string", "null"] }
        notificationStatus:
          type: string
          enum: [accepted, pending, failed, not_configured]
        createdAt: { type: string, format: date-time }
        assignedAt: { type: ["string", "null"], format: date-time }
        repliedAt: { type: ["string", "null"], format: date-time }
        closedAt: { type: ["string", "null"], format: date-time }
        updatedAt: { type: string, format: date-time }
...
    TravellerProfile:
      type: object
      required: [id, userId, displayName, homeBase, bio, createdAt, updatedAt]
      properties:
        id: { type: integer }
        userId: { type: string }
        displayName: { type: string }
        homeBase: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProfileUpdate:
      type: object
      properties:
        displayName: { type: string, minLength: 1, maxLength: 80 }
        homeBase: { type: ["string", "null"], maxLength: 120 }
        bio: { type: ["string", "null"], maxLength: 500 }
    SavedTrip:
      type: object
      required: [id, title, origin, destination, dateRange, travellerCount, budget, sustainabilityPriority, context, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        origin: { type: ["string", "null"] }
        destination: { type: ["string", "null"] }
        dateRange: { type: ["string", "null"] }
        travellerCount: { type: ["integer", "null"] }
        budget: { type: ["string", "null"] }
        sustainabilityPriority: { type: ["string", "null"] }
        context: { type: object, additionalProperties: true }
        status: { type: string, enum: [planning, ready, archived] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SavedTripInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        origin: { type: ["string", "null"] }
        destination: { type: ["string", "null"] }
        dateRange: { type: ["string", "null"] }
        travellerCount: { type: ["integer", "null"], minimum: 1, maximum: 20 }
        budget: { type: ["string", "null"] }
        sustainabilityPriority: { type: ["string", "null"] }
        context: { type: object, additionalProperties: true }
        status: { type: string, enum: [planning, ready, archived] }
    SavedTripUpdate:
      allOf:
        - $ref: "#/components/schemas/SavedTripInput"
    Booking:
      type: object
      required: [id, tripId, label, bookingType, providerReference, status, source, carbonKg, details, createdAt]
      properties:
        id: { type: integer }
        tripId: { type: ["integer", "null"] }
        label: { type: string }
        bookingType: { type: string, enum: [transport, stay, experience] }
        providerReference: { type: ["string", "null"] }
        status: { type: string, enum: [requested, demo, confirmed, cancelled] }
        source: { type: string, enum: [demo, live] }
        carbonKg: { type: ["number", "null"] }
        details: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
    BookingInput:
      type: object
      required: [label, bookingType]
      properties:
        tripId: { type: ["integer", "null"] }
        label: { type: string, minLength: 1, maxLength: 160 }
        bookingType: { type: string, enum: [transport, stay, experience] }
        providerReference: { type: ["string", "null"] }
        carbonKg: { type: ["number", "null"], minimum: 0 }
        details: { type: object, additionalProperties: true }
    RewardEvent:
      type: object
      required: [id, points, reason, evidence, category, createdAt]
      properties:
        id: { type: integer }
        points: { type: integer }
        reason: { type: string }
        evidence: { type: string }
        category: { type: string }
        createdAt: { type: string, format: date-time }
    RewardSummary:
      type: object
      required: [totalPoints, level, nextLevelPoints, events]
      properties:
        totalPoints: { type: integer }
        level: { type: string }
        nextLevelPoints: { type: integer }
        events:
          type: array
          items: { $ref: "#/components/schemas/RewardEvent" }
    ContactInput:
      type: object
      required: [name, email, message, consent]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        email: { type: string, format: email }
        message: { type: string, minLength: 1, maxLength: 2000 }
        topic: { type: ["string", "null"], maxLength: 80 }
        consent: { type: boolean }
    ContactResponse:
      type: object
      required: [id, status]
      properties:
        id: { type: integer }
        status: { type: string, enum: [received] }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer